requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.0

# Google Gemini (FREE!)
google-generativeai>=0.3.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Azure/MS Fabric (optional)
azure-identity>=1.15.0
azure-storage-file-datalake>=12.14.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
